// src/app/characters/dto/character.dto.ts
// Generated by https://quicktype.io

export interface Character {
    account_name: string;
    charId: number;
    char_name: string;
    level: number;
    maxHp: number;
    curHp: number;
    maxCp: number;
    curCp: number;
    maxMp: number;
    curMp: number;
    sex: number;
    exp: number;
    pvpkills: number;
    pkkills: number;
    clanid: number;
    race: number;
    classid: number;
    base_class: number;
    title?: string;
    title_color?: number;
    power_grade: number;
    nobless: number;
    vitality_points: number;
}